split up types into files
