+++ src/lib.rs
@@ -55,10 +55,6 @@ use dialoguer::Confirm;
 use dunce::canonicalize;
 use fs_err as fs;
 use futures::FutureExt;
-use metadata::{
-    BuildConfiguration, BuildSummary, Output, PackageIdentifier, PackagingSettings,
-    build_reindexed_channels,
-};
 use miette::{Context, IntoDiagnostic};
 pub use normalized_key::NormalizedKey;
 use opt::*;
@@ -80,10 +76,13 @@ use source_code::Source;
 use system_tools::SystemTools;
 use tool_configuration::{Configuration, ContinueOnFailure, SkipExisting, TestStrategy};
 use types::Directories;
+use types::{
+    BuildConfiguration, BuildSummary, PackageIdentifier, PackagingSettings,
+    build_reindexed_channels,
+};
 use variant_config::VariantConfig;
 
-use crate::metadata::Debug;
-use crate::metadata::PlatformWithVirtualPackages;
+use crate::metadata::{Debug, Output, PlatformWithVirtualPackages};
 
 /// Returns the recipe path.
 pub fn get_recipe_path(path: &Path) -> miette::Result<PathBuf> {
@@ -356,7 +355,7 @@ pub async fn get_build_output(
 
         let timestamp = chrono::Utc::now();
         let virtual_package_override = VirtualPackageOverrides::from_env();
-        let output = metadata::Output {
+        let output = Output {
             recipe: recipe.clone(),
             build_configuration: BuildConfiguration {
                 target_platform: discovered_output.target_platform,
@@ -823,7 +822,7 @@ pub async fn rebuild(
     let rendered_recipe =
         fs::read_to_string(temp_dir.join("rendered_recipe.yaml")).into_diagnostic()?;
 
-    let mut output: metadata::Output = serde_yaml::from_str(&rendered_recipe).into_diagnostic()?;
+    let mut output: Output = serde_yaml::from_str(&rendered_recipe).into_diagnostic()?;
 
     // set recipe dir to the temp folder
     output.build_configuration.directories.recipe_dir = temp_dir;
