+++ src/metadata.rs
@@ -1,620 +1,7 @@
 //! All the metadata that makes up a recipe file
-use std::{
-    borrow::Cow,
-    collections::BTreeMap,
-    fmt::{self, Display, Formatter},
-    io::Write,
-    iter,
-    path::{Path, PathBuf},
-    str::FromStr,
-    sync::{Arc, Mutex},
+pub use crate::types::{
+    BuildConfiguration, Debug, Output, PlatformWithVirtualPackages, build_reindexed_channels,
 };
-
-use chrono::{DateTime, Utc};
-use fs_err as fs;
-use indicatif::HumanBytes;
-use rattler_conda_types::{
-    Channel, ChannelUrl, GenericVirtualPackage, PackageName, Platform, RepoDataRecord,
-    VersionWithSource,
-    compression_level::CompressionLevel,
-    package::{ArchiveType, PathType, PathsEntry, PathsJson},
-};
-use rattler_index::{IndexFsConfig, index_fs};
-use rattler_repodata_gateway::SubdirSelection;
-use rattler_solve::{ChannelPriority, SolveStrategy};
-use rattler_virtual_packages::{
-    DetectVirtualPackageError, VirtualPackageOverrides, VirtualPackages,
-};
-use serde::{Deserialize, Deserializer, Serialize};
-use serde_json::Value;
-
-use crate::{
-    console_utils::github_integration_enabled,
-    hash::HashInfo,
-    normalized_key::NormalizedKey,
-    recipe::{
-        jinja::SelectorConfig,
-        parser::{Recipe, Source},
-        variable::Variable,
-    },
-    render::resolved_dependencies::FinalizedDependencies,
-    script::SandboxConfiguration,
-    system_tools::SystemTools,
-    tool_configuration,
-    types::Directories,
-};
-
-/// Default value for store recipe for backwards compatibility
-fn default_true() -> bool {
-    true
-}
-
-/// Settings when creating the package (compression etc.)
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct PackagingSettings {
-    /// The archive type, currently supported are `tar.bz2` and `conda`
-    pub archive_type: ArchiveType,
-    /// The compression level from 1-9 or -7-22 for `tar.bz2` and `conda`
-    /// archives
-    pub compression_level: i32,
-}
-
-impl PackagingSettings {
-    /// Create a new `PackagingSettings` from the command line arguments
-    /// and the selected archive type.
-    pub fn from_args(archive_type: ArchiveType, compression_level: CompressionLevel) -> Self {
-        let compression_level: i32 = match archive_type {
-            ArchiveType::TarBz2 => compression_level.to_bzip2_level().unwrap() as i32,
-            ArchiveType::Conda => compression_level.to_zstd_level().unwrap(),
-        };
-
-        Self {
-            archive_type,
-            compression_level,
-        }
-    }
-}
-
-/// Defines both a platform and the virtual packages that describe the
-/// capabilities of the platform.
-#[derive(Debug, Clone, Serialize)]
-pub struct PlatformWithVirtualPackages {
-    /// The platform
-    pub platform: Platform,
-
-    /// The virtual packages for the platform
-    pub virtual_packages: Vec<GenericVirtualPackage>,
-}
-
-impl PlatformWithVirtualPackages {
-    /// Returns the current platform and the virtual packages available on the
-    /// current system.
-    pub fn detect(overrides: &VirtualPackageOverrides) -> Result<Self, DetectVirtualPackageError> {
-        let platform = Platform::current();
-        Self::detect_for_platform(platform, overrides)
-    }
-
-    /// Detect the virtual packages for the given platform, filling in defaults where appropriate
-    pub fn detect_for_platform(
-        platform: Platform,
-        overrides: &VirtualPackageOverrides,
-    ) -> Result<Self, DetectVirtualPackageError> {
-        let virtual_packages = VirtualPackages::detect_for_platform(platform, overrides)?
-            .into_generic_virtual_packages()
-            .collect();
-        Ok(Self {
-            platform,
-            virtual_packages,
-        })
-    }
-}
-
-impl<'de> Deserialize<'de> for PlatformWithVirtualPackages {
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        #[derive(Deserialize)]
-        pub struct Object {
-            pub platform: Platform,
-            pub virtual_packages: Vec<GenericVirtualPackage>,
-        }
-
-        serde_untagged::UntaggedEnumVisitor::new()
-            .string(|s| {
-                Ok(Self {
-                    platform: Platform::from_str(s).map_err(serde::de::Error::custom)?,
-                    virtual_packages: vec![],
-                })
-            })
-            .map(|m| {
-                let object: Object = m.deserialize()?;
-                Ok(Self {
-                    platform: object.platform,
-                    virtual_packages: object.virtual_packages,
-                })
-            })
-            .deserialize(deserializer)
-    }
-}
-
-/// A newtype wrapper around a boolean indicating whether debug output is enabled
-#[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct Debug(bool);
-
-impl Debug {
-    /// Create a new Debug instance
-    pub fn new(debug: bool) -> Self {
-        Self(debug)
-    }
-
-    /// Returns true if debug output is enabled
-    pub fn is_enabled(&self) -> bool {
-        self.0
-    }
-}
-
-/// The configuration for a build of a package
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BuildConfiguration {
-    /// The target platform for the build
-    pub target_platform: Platform,
-    /// The host platform (usually target platform, but for `noarch` it's the
-    /// build platform)
-    pub host_platform: PlatformWithVirtualPackages,
-    /// The build platform (the platform that the build is running on)
-    pub build_platform: PlatformWithVirtualPackages,
-    /// The selected variant for this build
-    pub variant: BTreeMap<NormalizedKey, Variable>,
-    /// THe computed hash of the variant
-    pub hash: HashInfo,
-    /// The directories for the build (work, source, build, host, ...)
-    pub directories: Directories,
-    /// The channels to use when resolving environments
-    pub channels: Vec<ChannelUrl>,
-    /// The channel priority that is used to resolve dependencies
-    pub channel_priority: ChannelPriority,
-    /// The solve strategy to use when resolving dependencies
-    pub solve_strategy: SolveStrategy,
-    /// The timestamp to use for the build
-    pub timestamp: chrono::DateTime<chrono::Utc>,
-    /// All subpackages coming from this output or other outputs from the same
-    /// recipe
-    pub subpackages: BTreeMap<PackageName, PackageIdentifier>,
-    /// Package format (.tar.bz2 or .conda)
-    pub packaging_settings: PackagingSettings,
-    /// Whether to store the recipe and build instructions in the final package
-    /// or not
-    #[serde(skip_serializing, default = "default_true")]
-    pub store_recipe: bool,
-    /// Whether to set additional environment variables to force colors in the
-    /// build script or not
-    #[serde(skip_serializing, default = "default_true")]
-    pub force_colors: bool,
-
-    /// The configuration for the sandbox
-    #[serde(skip_serializing, default)]
-    pub sandbox_config: Option<SandboxConfiguration>,
-    /// Whether to enable debug output in build scripts
-    #[serde(skip_serializing, default)]
-    pub debug: Debug,
-    /// Exclude packages newer than this date from the solver
-    #[serde(skip_serializing, default)]
-    pub exclude_newer: Option<chrono::DateTime<chrono::Utc>>,
-}
-
-impl BuildConfiguration {
-    /// true if the build is cross-compiling
-    pub fn cross_compilation(&self) -> bool {
-        self.target_platform != self.build_platform.platform
-    }
-
-    /// Retrieve the sandbox configuration for this output
-    pub fn sandbox_config(&self) -> Option<&SandboxConfiguration> {
-        self.sandbox_config.as_ref()
-    }
-
-    /// Construct a `SelectorConfig` from the given `BuildConfiguration`
-    pub fn selector_config(&self) -> SelectorConfig {
-        SelectorConfig {
-            target_platform: self.target_platform,
-            host_platform: self.host_platform.platform,
-            build_platform: self.build_platform.platform,
-            variant: self.variant.clone(),
-            hash: Some(self.hash.clone()),
-            experimental: false,
-            allow_undefined: false,
-            recipe_path: None,
-        }
-    }
-}
-
-/// A package identifier
-#[derive(Serialize, Deserialize, Debug, Clone)]
-pub struct PackageIdentifier {
-    /// The name of the package
-    pub name: PackageName,
-    /// The version of the package
-    pub version: VersionWithSource,
-    /// The build string of the package
-    pub build_string: String,
-}
-
-/// The summary of a build
-#[derive(Debug, Clone, Default)]
-pub struct BuildSummary {
-    /// The start time of the build
-    pub build_start: Option<DateTime<Utc>>,
-    /// The end time of the build
-    pub build_end: Option<DateTime<Utc>>,
-
-    /// The path to the artifact
-    pub artifact: Option<PathBuf>,
-    /// Any warnings that were recorded during the build
-    pub warnings: Vec<String>,
-    /// The paths that are packaged in the artifact
-    pub paths: Option<PathsJson>,
-    ///  Whether the build was successful or not
-    pub failed: bool,
-}
-
-/// A output. This is the central element that is passed to the `run_build`
-/// function and fully specifies all the options and settings to run the build.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct Output {
-    /// The rendered recipe that is used to build this output
-    pub recipe: Recipe,
-    /// The build configuration for this output (e.g. target_platform, channels,
-    /// and other settings)
-    pub build_configuration: BuildConfiguration,
-    /// The finalized dependencies for this output. If this is `None`, the
-    /// dependencies have not been resolved yet. During the `run_build`
-    /// functions, the dependencies are resolved and this field is filled.
-    pub finalized_dependencies: Option<FinalizedDependencies>,
-    /// The finalized sources for this output. Contain the exact git hashes for
-    /// the sources that are used to build this output.
-    pub finalized_sources: Option<Vec<Source>>,
-
-    /// The finalized dependencies from the cache (if there is a cache
-    /// instruction)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub finalized_cache_dependencies: Option<FinalizedDependencies>,
-    /// The finalized sources from the cache (if there is a cache instruction)
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub finalized_cache_sources: Option<Vec<Source>>,
-
-    /// Summary of the build
-    #[serde(skip)]
-    pub build_summary: Arc<Mutex<BuildSummary>>,
-    /// The system tools that are used to build this output
-    pub system_tools: SystemTools,
-    /// Some extra metadata that should be recorded additionally in about.json
-    /// Usually it is used during the CI build to record link to the CI job
-    /// that created this artifact
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub extra_meta: Option<BTreeMap<String, Value>>,
-}
-
-impl Output {
-    /// The name of the package
-    pub fn name(&self) -> &PackageName {
-        self.recipe.package().name()
-    }
-
-    /// The version of the package
-    pub fn version(&self) -> &VersionWithSource {
-        self.recipe.package().version()
-    }
-
-    /// The build string is either the build string from the recipe or computed
-    /// from the hash and build number.
-    pub fn build_string(&self) -> Cow<'_, str> {
-        self.recipe
-            .build()
-            .string
-            .as_resolved()
-            .expect("Build string is not resolved")
-            .into()
-    }
-
-    /// retrieve an identifier for this output ({name}-{version}-{build_string})
-    pub fn identifier(&self) -> String {
-        format!(
-            "{}-{}-{}",
-            self.name().as_normalized(),
-            self.version(),
-            &self.build_string()
-        )
-    }
-
-    /// Record a warning during the build
-    pub fn record_warning(&self, warning: &str) {
-        self.build_summary
-            .lock()
-            .unwrap()
-            .warnings
-            .push(warning.to_string());
-    }
-
-    /// Record the start of the build
-    pub fn record_build_start(&self) {
-        self.build_summary.lock().unwrap().build_start = Some(chrono::Utc::now());
-    }
-
-    /// Record the artifact that was created during the build
-    pub fn record_artifact(&self, artifact: &Path, paths: &PathsJson) {
-        let mut summary = self.build_summary.lock().unwrap();
-        summary.artifact = Some(artifact.to_path_buf());
-        summary.paths = Some(paths.clone());
-    }
-
-    /// Record the end of the build
-    pub fn record_build_end(&self) {
-        let mut summary = self.build_summary.lock().unwrap();
-        summary.build_end = Some(chrono::Utc::now());
-    }
-
-    /// Shorthand to retrieve the variant configuration for this output
-    pub fn variant(&self) -> &BTreeMap<NormalizedKey, Variable> {
-        &self.build_configuration.variant
-    }
-
-    /// Shorthand to retrieve the host prefix for this output
-    pub fn prefix(&self) -> &Path {
-        &self.build_configuration.directories.host_prefix
-    }
-
-    /// Shorthand to retrieve the build prefix for this output
-    pub fn build_prefix(&self) -> &Path {
-        &self.build_configuration.directories.build_prefix
-    }
-
-    /// Shorthand to retrieve the target platform for this output
-    pub fn target_platform(&self) -> &Platform {
-        &self.build_configuration.target_platform
-    }
-
-    /// Shorthand to retrieve the target platform for this output
-    pub fn host_platform(&self) -> &PlatformWithVirtualPackages {
-        &self.build_configuration.host_platform
-    }
-
-    /// Search for the resolved package with the given name in the host prefix
-    /// Returns a tuple of the package and a boolean indicating whether the
-    /// package is directly requested
-    pub fn find_resolved_package(&self, name: &str) -> Option<(&RepoDataRecord, bool)> {
-        let host = self.finalized_dependencies.as_ref()?.host.as_ref()?;
-        let record = host
-            .resolved
-            .iter()
-            .find(|p| p.package_record.name.as_normalized() == name);
-
-        let is_requested = host.specs.iter().any(|s| {
-            s.spec()
-                .name
-                .as_ref()
-                .map(|n| n.as_normalized() == name)
-                .unwrap_or(false)
-        });
-
-        record.map(|r| (r, is_requested))
-    }
-
-    /// Print a nice summary of the build
-    pub fn log_build_summary(&self) -> Result<(), std::io::Error> {
-        let summary = self.build_summary.lock().unwrap();
-        let identifier = self.identifier();
-        let span = tracing::info_span!("Build summary for", recipe = identifier);
-        let _enter = span.enter();
-
-        if let Some(artifact) = &summary.artifact {
-            let bytes = HumanBytes(fs::metadata(artifact).map(|m| m.len()).unwrap_or(0));
-            tracing::info!("Artifact: {} ({})", artifact.display(), bytes);
-        } else {
-            tracing::info!("No artifact was created");
-        }
-        tracing::info!("{}", self);
-
-        if !summary.warnings.is_empty() {
-            tracing::warn!("Warnings:");
-            for warning in &summary.warnings {
-                tracing::warn!("{}", warning);
-            }
-        }
-
-        if let Ok(github_summary) = std::env::var("GITHUB_STEP_SUMMARY") {
-            if !github_integration_enabled() {
-                return Ok(());
-            }
-            // append to the summary file
-            let mut summary_file = fs::OpenOptions::new()
-                .append(true)
-                .create(true)
-                .open(github_summary)?;
-
-            writeln!(summary_file, "### Build summary for {}", identifier)?;
-            if let Some(article) = &summary.artifact {
-                let bytes = HumanBytes(fs::metadata(article).map(|m| m.len()).unwrap_or(0));
-                writeln!(
-                    summary_file,
-                    "**Artifact**: {} ({})",
-                    article.display(),
-                    bytes
-                )?;
-            } else {
-                writeln!(summary_file, "**No artifact was created**")?;
-            }
-
-            if let Some(paths) = &summary.paths {
-                if paths.paths.is_empty() {
-                    writeln!(summary_file, "Included files: **No files included**")?;
-                } else {
-                    /// Github detail expander
-                    fn format_entry(entry: &PathsEntry) -> String {
-                        let mut extra_info = Vec::new();
-                        if entry.prefix_placeholder.is_some() {
-                            extra_info.push("contains prefix");
-                        }
-                        if entry.no_link {
-                            extra_info.push("no link");
-                        }
-                        match entry.path_type {
-                            PathType::SoftLink => extra_info.push("soft link"),
-                            // skip default
-                            PathType::HardLink => {}
-                            PathType::Directory => extra_info.push("directory"),
-                        }
-                        let bytes = entry.size_in_bytes.unwrap_or(0);
-
-                        format!(
-                            "| `{}` | {} | {} |",
-                            entry.relative_path.to_string_lossy(),
-                            HumanBytes(bytes),
-                            extra_info.join(", ")
-                        )
-                    }
-
-                    writeln!(summary_file, "<details>")?;
-                    writeln!(
-                        summary_file,
-                        "<summary>Included files ({} files)</summary>\n",
-                        paths.paths.len()
-                    )?;
-                    writeln!(summary_file, "| Path | Size | Extra info |")?;
-                    writeln!(summary_file, "| --- | --- | --- |")?;
-                    for path in &paths.paths {
-                        writeln!(summary_file, "{}", format_entry(path))?;
-                    }
-                    writeln!(summary_file, "\n</details>\n")?;
-                }
-            }
-
-            if !summary.warnings.is_empty() {
-                writeln!(summary_file, "> [!WARNING]")?;
-                writeln!(summary_file, "> **Warnings during build:**\n>")?;
-                for warning in &summary.warnings {
-                    writeln!(summary_file, "> - {}", warning)?;
-                }
-                writeln!(summary_file)?;
-            }
-
-            writeln!(
-                summary_file,
-                "<details><summary>Resolved dependencies</summary>\n\n{}\n</details>\n",
-                self.format_as_markdown()
-            )?;
-        }
-        Ok(())
-    }
-}
-
-impl Output {
-    /// Format the output as a markdown table
-    pub fn format_as_markdown(&self) -> String {
-        let mut output = String::new();
-        self.format_table_with_option(&mut output, comfy_table::presets::ASCII_MARKDOWN, true)
-            .expect("Could not format table");
-        output
-    }
-
-    fn format_table_with_option(
-        &self,
-        f: &mut impl fmt::Write,
-        table_format: &str,
-        long: bool,
-    ) -> std::fmt::Result {
-        let template = || -> comfy_table::Table {
-            let mut table = comfy_table::Table::new();
-            if table_format == comfy_table::presets::UTF8_FULL {
-                table
-                    .load_preset(comfy_table::presets::UTF8_FULL_CONDENSED)
-                    .apply_modifier(comfy_table::modifiers::UTF8_ROUND_CORNERS);
-            } else {
-                table.load_preset(table_format);
-            }
-            table
-        };
-
-        writeln!(f, "Variant configuration (hash: {}):", self.build_string())?;
-        let mut table = template();
-        if table_format != comfy_table::presets::UTF8_FULL {
-            table.set_header(["Key", "Value"]);
-        }
-        self.build_configuration.variant.iter().for_each(|(k, v)| {
-            table.add_row([k.normalize(), format!("{:?}", v)]);
-        });
-        writeln!(f, "{}\n", table)?;
-
-        if let Some(finalized_dependencies) = &self.finalized_dependencies {
-            if let Some(build) = &finalized_dependencies.build {
-                writeln!(f, "Build dependencies:")?;
-                writeln!(f, "{}\n", build.to_table(template(), long))?;
-            }
-
-            if let Some(host) = &finalized_dependencies.host {
-                writeln!(f, "Host dependencies:")?;
-                writeln!(f, "{}\n", host.to_table(template(), long))?;
-            }
-
-            if !finalized_dependencies.run.depends.is_empty() {
-                writeln!(f, "Run dependencies:")?;
-                writeln!(
-                    f,
-                    "{}\n",
-                    finalized_dependencies.run.to_table(template(), long)
-                )?;
-            }
-        }
-
-        Ok(())
-    }
-}
-
-impl Display for Output {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        self.format_table_with_option(f, comfy_table::presets::UTF8_FULL, false)
-    }
-}
-
-/// Builds the channel list and reindexes the output channel.
-pub async fn build_reindexed_channels(
-    build_configuration: &BuildConfiguration,
-    tool_configuration: &tool_configuration::Configuration,
-) -> Result<Vec<ChannelUrl>, std::io::Error> {
-    let output_dir = &build_configuration.directories.output_dir;
-    let output_channel = Channel::from_directory(output_dir);
-
-    // Clear the repodata gateway of any cached values for the output channel.
-    tool_configuration.repodata_gateway.clear_repodata_cache(
-        &output_channel,
-        SubdirSelection::Some(
-            [build_configuration.target_platform]
-                .iter()
-                .map(ToString::to_string)
-                .collect(),
-        ),
-    );
-
-    let index_config = IndexFsConfig {
-        channel: output_dir.clone(),
-        target_platform: Some(build_configuration.target_platform),
-        repodata_patch: None,
-        write_zst: false,
-        write_shards: false,
-        force: false,
-        max_parallel: num_cpus::get_physical(),
-        multi_progress: None,
-    };
-
-    // Reindex the output channel from the files on disk
-    index_fs(index_config)
-        .await
-        .map_err(|e| std::io::Error::new(std::io::ErrorKind::Other, e))?;
-
-    Ok(iter::once(output_channel.base_url)
-        .chain(build_configuration.channels.iter().cloned())
-        .collect())
-}
 
 #[cfg(test)]
 mod test {
