+++ src/source/mod.rs
@@ -6,7 +6,6 @@ use std::{
 };
 
 use crate::{
-    Output,
     recipe::parser::{GitRev, GitSource, Source},
     source::{
         checksum::Checksum,
@@ -14,7 +13,7 @@ use crate::{
     },
     system_tools::ToolError,
     tool_configuration,
-    types::Directories,
+    types::{Directories, Output},
 };
 
 use fs_err as fs;
